+++ plugins/announcements-backend/CHANGELOG.md
@@ -1,4 +1,28 @@
-# @k-phoen/backstage-plugin-announcements-backend
+# @procore-oss/backstage-plugin-announcements-backend
+
+## 0.3.5
+
+### Patch Changes
+
+- 90a19ec: upgrade backstage to v1.18.0
+- c3c379d: bump backstage to v1.16.0
+- Updated dependencies [90a19ec]
+- Updated dependencies [c3c379d]
+  - @procore-oss/backstage-plugin-announcements-common@0.1.2
+
+## 0.3.4
+
+### Patch Changes
+
+- 57792bf: Fixed the correct endpoint call by AnnouncementCollatorApi
+
+## 0.3.3
+
+### Patch Changes
+
+- 4f7a351: bump all packages
+- Updated dependencies [4f7a351]
+  - @procore-oss/backstage-plugin-announcements-common@0.1.1
 
 ## 0.2.0
 
@@ -13,7 +37,7 @@
 - b8c5c87: Paginate results in the announcements page
 - Updated dependencies [793d5b9]
 - Updated dependencies [ab3813f]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.7
+  - @procore-oss/backstage-plugin-announcements-common@0.0.7
 
 ## 0.1.4
 
@@ -21,7 +45,7 @@
 
 - 7d25e84: Bump Backstage-related dependencies
 - Updated dependencies [7d25e84]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.6
+  - @procore-oss/backstage-plugin-announcements-common@0.0.6
 
 ## 0.1.3
 
@@ -31,21 +55,21 @@
 - 56d5e6d: Try and setup release pipeline to replace 'workspace:\*' version constraints
 - Updated dependencies [0e1d000]
 - Updated dependencies [56d5e6d]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.5
+  - @procore-oss/backstage-plugin-announcements-common@0.0.5
 
 ## 0.1.2
 
 ### Patch Changes
 
 - Updated dependencies [9bdc37d]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.4
+  - @procore-oss/backstage-plugin-announcements-common@0.0.4
 
 ## 0.1.1
 
 ### Patch Changes
 
 - Updated dependencies [6c9bf32]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.3
+  - @procore-oss/backstage-plugin-announcements-common@0.0.3
 
 ## 0.1.0
 
@@ -58,7 +82,7 @@
 - 0c12eea: Bump Backstage dependencies
 - Updated dependencies [062aca5]
 - Updated dependencies [0c12eea]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.2
+  - @procore-oss/backstage-plugin-announcements-common@0.0.2
 
 ## 0.0.5