+++ plugins/announcements/CHANGELOG.md
@@ -1,4 +1,22 @@
-# @k-phoen/backstage-plugin-announcements
+# @procore-oss/backstage-plugin-announcements
+
+## 0.2.3
+
+### Patch Changes
+
+- 90a19ec: upgrade backstage to v1.18.0
+- c3c379d: bump backstage to v1.16.0
+- Updated dependencies [90a19ec]
+- Updated dependencies [c3c379d]
+  - @procore-oss/backstage-plugin-announcements-common@0.1.2
+
+## 0.2.2
+
+### Patch Changes
+
+- 4f7a351: bump all packages
+- Updated dependencies [4f7a351]
+  - @procore-oss/backstage-plugin-announcements-common@0.1.1
 
 ## 0.1.0
 
@@ -17,7 +35,7 @@
 - cca69ec: Prevent the submission of empty announcements
 - Updated dependencies [793d5b9]
 - Updated dependencies [ab3813f]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.7
+  - @procore-oss/backstage-plugin-announcements-common@0.0.7
 
 ## 0.0.14
 
@@ -44,7 +62,7 @@
 - 00753b7: Fix routes definition
 - 60babbc: Make the themeId configurable on all pages
 - Updated dependencies [7d25e84]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.6
+  - @procore-oss/backstage-plugin-announcements-common@0.0.6
 
 ## 0.0.11
 
@@ -54,21 +72,21 @@
 - 56d5e6d: Try and setup release pipeline to replace 'workspace:\*' version constraints
 - Updated dependencies [0e1d000]
 - Updated dependencies [56d5e6d]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.5
+  - @procore-oss/backstage-plugin-announcements-common@0.0.5
 
 ## 0.0.10
 
 ### Patch Changes
 
 - Updated dependencies [9bdc37d]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.4
+  - @procore-oss/backstage-plugin-announcements-common@0.0.4
 
 ## 0.0.9
 
 ### Patch Changes
 
 - Updated dependencies [6c9bf32]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.3
+  - @procore-oss/backstage-plugin-announcements-common@0.0.3
 
 ## 0.0.8
 
@@ -77,7 +95,7 @@
 - 0c12eea: Bump Backstage dependencies
 - Updated dependencies [062aca5]
 - Updated dependencies [0c12eea]
-  - @k-phoen/backstage-plugin-announcements-common@0.0.2
+  - @procore-oss/backstage-plugin-announcements-common@0.0.2
 
 ## 0.0.7